Added support for Windows builds of Xamarin.Mac
