+++ msbuild/Xamarin.Mac.Tasks.Core/Tasks/XamMacArch.cs
